% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierSIR.R
\name{outlierSIR}
\alias{outlierSIR}
\title{Check for outliers in SIR models.}
\usage{
outlierSIR(X, y, method = c("BOOT", "TTR", "MONO"), kernel = "normal",
  bandwidth = NULL, H = 5, nb.replications = 1000, pourcent = 0.1,
  plotBandwidthEst = F, NCORES = 1)
}
\arguments{
\item{X}{A matrix nxp describinf n observations through p varaibles.}

\item{y}{A vector of \strong{n} observations corresponding to the n rows of \emph{X}.}

\item{method}{Which method to be performed. One of \emph{BOOT} for bootstrap, \emph{TTR} for train-test replication and \emph{MONO} for mono-model testing.}

\item{kernel}{The kernel to be used, one of "normal" or "box". Default is "normal". Can be abbreviated. See \code{\link{ksmooth}}.}

\item{bandwidth}{A real indicating the width of the kernel. If NULL, automatic kernel width tuning is used. Default is "NULL".}

\item{H}{A positive real value. The number of slices in the SIR model. Default is 5.}

\item{nb.replications}{An interger giving the number of repliccations for \emph{BOOT} and \emph{TTR} methods. Default is "1000".}

\item{pourcent}{A real between 0 and 1. The proportion of observations to keep in the test dataset. Default is "0.1".}

\item{plotBandwidthEst}{An integer. Wether or not to plot the EDR direction with the kernel estimation. Default is "FALSE".}

\item{NCORES}{An integer. Fiwing the number of cores to be used in the sampling computations. Only BOOT method has been parallelized due to its computation time.}
}
\value{
A list containing indices of outliers in the case of the 3 methods and also the extreme observations in the case of \emph{BOOT}.
}
\description{
This function takes a matrix \eqn{X} defining the same \eqn{n} individuals and a vector \eqn{y} defining also those individuals.
}
\examples{
data(ozone)
# MONO method
output <- outlierSIR(ozone[,-c(1,12,13)],ozone[,1],method="MONO")
# TTR method
# output <- outlierSIR(ozone[,-c(1,12,13)],ozone[,1],method="TTR",
# plotBandwidthEst = FALSE,kernel="normal",nb.replications = 1000)
# BOOT method, using 5 cores of computation
# output <- outlierSIR(ozone[,-c(1,12,13)],ozone[,1],method="BOOT",
# plotBandwidthEst = TRUE,kernel="normal",nb.replications = 5000,NCORES=5)
}
\references{
{
  \insertAllCited{}
}
}
